save of interaction matrix crashing

+++ src/calibration_functions.py
@@ -318,8 +318,8 @@ def create_response_matrix(
     push_pull_images = pp_sum / n_runs
 
     # Compute response matrices from the averaged push-pull images
-    response_matrix_full = compute_response_matrix(push_pull_images)
-    response_matrix_filtered = compute_response_matrix(push_pull_images, mask)
+    response_matrix_full = compute_response_matrix(push_pull_images).astype(np.float32)
+    response_matrix_filtered = compute_response_matrix(push_pull_images, mask).astype(np.float32)
 
     # Define output filenames
     pull_filename     = f'binned_processed_response_cube_KL2PWFS_only_pull_nact_{nact}_amp_{phase_amp}_3s_pyr.fits'
